class1/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
tqdm>=4.65.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pillow>=9.5.0
requests>=2.31.0
gdown>=5.0.0
kagglehub[pandas-datasets]
seaborn
ipykernel